% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_covariates.R
\name{estimate_beta_from_groups}
\alias{estimate_beta_from_groups}
\title{Estimate beta from group-level ATTs (internal)}
\usage{
estimate_beta_from_groups(df, x_group, conditional_model)
}
\description{
Estimate beta from group-level ATTs (internal)
}
\keyword{internal}
