% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_fixest.R
\name{extract_fixest_eif}
\alias{extract_fixest_eif}
\title{Extract influence functions from fixest object}
\usage{
extract_fixest_eif(x, sunab_idx, n)
}
\arguments{
\item{x}{fixest object}

\item{sunab_idx}{Indices of sunab coefficients}

\item{n}{Sample size}
}
\value{
List of EIF vectors or NULL if unavailable
}
\description{
Attempts to extract influence functions from a fixest object for variance
propagation. Returns NULL if not available.
}
\keyword{internal}
