% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_covariates.R
\name{integrate_finite_population}
\alias{integrate_finite_population}
\title{Integrate over finite-population target sample (internal)}
\usage{
integrate_finite_population(
  gt_object,
  conditional_model,
  beta,
  x_target,
  weights
)
}
\description{
Integrate over finite-population target sample (internal)
}
\keyword{internal}
