% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_covariates.R
\name{integrate_monte_carlo}
\alias{integrate_monte_carlo}
\title{Integrate via Monte Carlo sampling (internal)}
\usage{
integrate_monte_carlo(gt_object, conditional_model, beta, sampler, n_mc)
}
\description{
Integrate via Monte Carlo sampling (internal)
}
\keyword{internal}
