% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_cate.R
\name{new_cate_integration}
\alias{new_cate_integration}
\title{Construct a cate_integration object}
\usage{
new_cate_integration(estimate, phi, inf, w, design, form, n, n_eff)
}
\arguments{
\item{estimate}{Scalar FATT estimate.}

\item{phi}{Length-n transport EIF vector.}

\item{inf}{Output of \code{\link[=compute_variance]{compute_variance()}} (list with var, se, ci, level).}

\item{w}{Length-n transport weights.}

\item{design}{Design string.}

\item{form}{\code{"target_index"} or \code{"density_ratio"}.}

\item{n}{Source sample size.}

\item{n_eff}{Effective sample size.}
}
\value{
Object of class \code{cate_integration}.
}
\description{
Construct a cate_integration object
}
\keyword{internal}
