% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_cate.R
\name{print.cate_integration}
\alias{print.cate_integration}
\title{Print a cate_integration object}
\usage{
\method{print}{cate_integration}(x, ...)
}
\arguments{
\item{x}{A \code{cate_integration} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print a cate_integration object
}
