% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_extrapolate_ATT.R
\name{print.cv_extrapolate}
\alias{print.cv_extrapolate}
\title{Print method for cv_extrapolate objects}
\usage{
\method{print}{cv_extrapolate}(x, ...)
}
\arguments{
\item{x}{A \code{cv_extrapolate} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cv_extrapolate objects
}
