% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_models.R
\name{print.extrap_object_averaged}
\alias{print.extrap_object_averaged}
\title{Print method for extrap_object_averaged}
\usage{
\method{print}{extrap_object_averaged}(x, ...)
}
\arguments{
\item{x}{An \code{extrap_object_averaged} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for extrap_object_averaged
}
