% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_extrapolate_ATT.R
\name{summary.cv_extrapolate}
\alias{summary.cv_extrapolate}
\title{Summary method for cv_extrapolate objects}
\usage{
\method{summary}{cv_extrapolate}(object, ...)
}
\arguments{
\item{object}{A \code{cv_extrapolate} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for cv_extrapolate objects
}
