% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_confidence_level}
\alias{validate_confidence_level}
\title{Validate confidence level}
\usage{
validate_confidence_level(level, name = "level")
}
\arguments{
\item{level}{Confidence level scalar}

\item{name}{Variable name for error messages}
}
\description{
Validate confidence level
}
\keyword{internal}
