% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_eif_list}
\alias{validate_eif_list}
\title{Validate EIF list structure}
\usage{
validate_eif_list(phi, n, name = "phi")
}
\arguments{
\item{phi}{List of EIF vectors}

\item{n}{Expected length of each vector}

\item{name}{Variable name for error messages}
}
\description{
Validate EIF list structure
}
\keyword{internal}
