% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_group_weights}
\alias{validate_group_weights}
\title{Validate group weights}
\usage{
validate_group_weights(omega, n_groups, name = "omega", warn_sum = TRUE)
}
\arguments{
\item{omega}{Numeric vector of group weights}

\item{n_groups}{Expected number of groups}

\item{name}{Variable name for error messages}

\item{warn_sum}{Whether to warn if weights don't sum to 1}
}
\description{
Validate group weights
}
\keyword{internal}
