% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_gt_object}
\alias{validate_gt_object}
\title{Validate gt_object structure}
\usage{
validate_gt_object(gt_obj, name = "gt_object")
}
\arguments{
\item{gt_obj}{Object to validate}

\item{name}{Variable name for error messages}
}
\description{
Validate gt_object structure
}
\keyword{internal}
