% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_lengths_match}
\alias{validate_lengths_match}
\title{Validate lengths match}
\usage{
validate_lengths_match(x, y, name_x = "x", name_y = "y")
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}

\item{name_x}{Name of first vector}

\item{name_y}{Name of second vector}
}
\description{
Validate lengths match
}
\keyword{internal}
