% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_model_specs}
\alias{validate_model_specs}
\title{Validate model specification list}
\usage{
validate_model_specs(model_specs, name = "model_specs")
}
\arguments{
\item{model_specs}{Named list of model specifications}

\item{name}{Variable name for error messages}
}
\description{
Validate model specification list
}
\keyword{internal}
