% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_numeric_vector}
\alias{validate_numeric_vector}
\title{Validate numeric vector}
\usage{
validate_numeric_vector(x, name = "x", allow_na = FALSE)
}
\arguments{
\item{x}{Vector to validate}

\item{name}{Variable name for error messages}

\item{allow_na}{Whether to allow NA values}
}
\description{
Validate numeric vector
}
\keyword{internal}
