% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_scalar}
\alias{validate_scalar}
\title{Validate scalar numeric}
\usage{
validate_scalar(x, name = "x")
}
\arguments{
\item{x}{Value to validate}

\item{name}{Variable name for error messages}
}
\description{
Validate scalar numeric
}
\keyword{internal}
